+++ docs/pages.jl
@@ -5,7 +5,8 @@ getting_started_pages = Any[
         "Analytical CDF solutions" => "getting-started/tutorials/analytical-primarycensored-cdfs.md",
         "Exponentially tilted primary events" => "getting-started/tutorials/exponentially-tilted-primary-events.md",
         "Automatic differentiation backends" => "getting-started/tutorials/ad-backends.md",
-        "Fitting with Turing.jl" => "getting-started/tutorials/fitting-with-turing.md"
+        "Fitting with Turing.jl" => "getting-started/tutorials/fitting-with-turing.md",
+        "Marginal and latent forms" => "getting-started/tutorials/marginal-latent-equivalence.md"
     ],
     "Julia" => "getting-started/julia.md",
     "FAQ" => "getting-started/faq.md"